# docs/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: SEIP API
  version: 2.4.0
  summary: School Evidence Intelligence Platform — core API (evidence workflow + PA evaluation + reports/AI + PDF)
  description: |
    Contract v2.4 (CCR-009 cleanup) — list evidence carries aggregate `scan_status`;
    assignment detail carries `framework_version_id`. v2.3 (CCR-008) session
    lifecycle; v2.2 PDF; core remains evidence, PA evaluation, reports, AI suggest.

    **Business rules that bind every implementation**
    - Tenancy: every resource is school-scoped; cross-school access returns RES-001 (indistinguishable from not-found).
    - Taxonomy (frameworks/domains/indicators/weights) is versioned seed data — never hard-coded.
    - Scoring grain is per-evaluator (3 committee members); overall pass requires **every** evaluator ≥ 70% and the workload gate met.
    - Rounds are configurable 1..n per cycle; no endpoint assumes exactly 2.
    - Reports cite evidence by reference only (ReportSectionRef) — never embed file bytes.
    - PDF is generated from structured payload (not a pixel-perfect ก.ค.ศ. plate); suitable for draft/print review.
    - AI mapping suggestions always land as mapping_source=ai_suggested + status=suggested; human confirm required.
    - Error bodies follow error-codes.yaml; codes are stable API surface.

    **Deferred** — see `x-deferred`.
  license:
    name: Proprietary — SEIP project internal
    url: https://github.com/sooksun/aischool
  contact:
...
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
...
paths:
...
  /evidence:
    get:
      operationId: listEvidence
      summary: List evidence (school-scoped; teachers see their own by default)
      tags: [evidence]
      parameters:
        - name: owner_personnel_id
          in: query
          schema: { type: string, format: uuid }
        - name: category_id
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/EvidenceStatus' }
        - name: mapped_to_indicator_id
          in: query
          description: Only evidence having an active mapping to this indicator
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Paged evidence list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvidencePage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createEvidence
      summary: Create a logical evidence item (metadata only; files attach via two-phase upload)
      tags: [evidence]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EvidenceCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Evidence' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /evidence/{evidenceId}:
    get:
      operationId: getEvidence
      summary: Evidence detail with files and mappings
      tags: [evidence]
      parameters:
        - $ref: '#/components/parameters/EvidenceId'
      responses:
        '200':
          description: Evidence detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvidenceDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateEvidence
      summary: Update evidence metadata (owner or school admin)
      tags: [evidence]
      parameters:
        - $ref: '#/components/parameters/EvidenceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EvidencePatch' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Evidence' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      operationId: deleteEvidence
      summary: Soft-delete evidence (audited; storage GC is async policy)
      tags: [evidence]
      parameters:
        - $ref: '#/components/parameters/EvidenceId'
      responses:
        '204':
          description: Soft-deleted
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /evidence/{evidenceId}/files/initiate:
    post:
      operationId: initiateFileUpload
      summary: 'Phase 1: validate metadata and obtain an upload target'
      description: |
        Validates content type / size / (video) duration against the evidence category
        (UPL-001..003) **before** bytes move. Returns a provider-agnostic upload target
        (presigned URL or equivalent). Client uploads bytes directly to storage,
        then calls the complete endpoint.
      tags: [evidence]
      parameters:
        - $ref: '#/components/parameters/EvidenceId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FileUploadInitiate' }
      responses:
        '201':
          description: Upload target issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileUploadTarget' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /evidence/{evidenceId}/files/{fileId}/complete:
    post:
      operationId: completeFileUpload
      summary: 'Phase 2: confirm bytes uploaded; register file and enqueue virus scan'
      description: Emits `evidence.file.registered`; scan result arrives via `evidence.file.scan_completed` (events.yaml). File is served only when scan_status=clean (UPL-006).
      tags: [evidence]
      parameters:
        - $ref: '#/components/parameters/EvidenceId'
        - name: fileId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FileUploadComplete' }
      responses:
        '200':
          description: File registered (scan pending)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvidenceFile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    EvidenceId:
      name: evidenceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    Unauthorized:
      description: AUTH-* — missing/expired/invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: PERM-* — role or tenancy denial
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: RES-001 — not found or cross-tenant (indistinguishable by design)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    UnprocessableEntity:
      description: VAL/UPL/MAP/SCORE/CYCLE-* — semantically invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: RES-002/MAP-001/UPL-004/CYCLE-002/SCORE-002 — state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: Stable code from error-codes.yaml }
        message: { type: string }
        details:
          type: array
          items:
            type: object
            required: [field, issue]
            properties:
              field: { type: string }
              issue: { type: string }
        request_id: { type: string, format: uuid }
...
    EvidenceStatus:
      type: string
      enum: [draft, active, archived, rejected]
    MappingStatus:
      type: string
      enum: [suggested, confirmed, rejected, revoked]
    MappingSource:
      type: string
      enum: [human, ai_suggested]
      description: ai_suggested set only by suggestMappings (ADR-0007 local_heuristic)
...
    ScanStatus:
      type: string
      enum: [pending, clean, blocked]
...
    Evidence:
      type: object
      required: [id, school_id, owner_personnel_id, category_id, title, status, created_at]
      description: |
        Lifecycle (CCR-002): created as `draft`; the server promotes to `active` when the
        first file upload completes successfully. Clients never set draft→active directly;
        EvidencePatch.status is for archive/reject transitions only.
      properties:
        id: { type: string, format: uuid }
        school_id: { type: string, format: uuid }
        owner_personnel_id: { type: string, format: uuid }
        uploaded_by_user_id: { type: string, format: uuid }
        category_id: { type: string, format: uuid }
        title: { type: string }
        description: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/EvidenceStatus' }
        captured_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
        scan_status:
          description: |
            Aggregate of attached files (CCR-009). null when there are no files.
            Worst-of order: pending > blocked > clean. Present on list and detail so
            clients never need N+1 detail fetches for list badges.
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/ScanStatus'
    EvidenceCreate:
      type: object
      required: [category_id, title]
      properties:
        owner_personnel_id:
          type: [string, 'null']
          format: uuid
          description: Defaults to the caller's own personnel profile; admins may set another owner
        category_id: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 300 }
        description: { type: [string, 'null'], maxLength: 2000 }
        captured_at: { type: [string, 'null'], format: date-time }
    EvidencePatch:
      type: object
      minProperties: 1
      properties:
        title: { type: string, minLength: 1, maxLength: 300 }
        description: { type: [string, 'null'], maxLength: 2000 }
        status: { $ref: '#/components/schemas/EvidenceStatus' }
        captured_at: { type: [string, 'null'], format: date-time }
    EvidenceDetail:
      allOf:
        - $ref: '#/components/schemas/Evidence'
        - type: object
          required: [files, mappings]
          properties:
            files:
              type: array
              items: { $ref: '#/components/schemas/EvidenceFile' }
            mappings:
              type: array
              items: { $ref: '#/components/schemas/Mapping' }
    EvidenceFile:
      type: object
      required: [id, evidence_id, content_type, byte_size, scan_status, uploaded_at]
      properties:
        id: { type: string, format: uuid }
        evidence_id: { type: string, format: uuid }
        content_type: { type: string }
        byte_size: { type: integer, minimum: 0 }
        checksum_sha256: { type: string }
        duration_seconds: { type: [integer, 'null'] }
        original_filename: { type: string }
        scan_status: { $ref: '#/components/schemas/ScanStatus' }
        download_url:
          type: [string, 'null']
          description: Short-lived read URL; null until scan_status=clean
        uploaded_at: { type: string, format: date-time }
    FileUploadInitiate:
      type: object
      required: [content_type, byte_size, checksum_sha256, original_filename]
      properties:
        content_type: { type: string }
        byte_size: { type: integer, minimum: 1 }
        checksum_sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        original_filename: { type: string, maxLength: 255 }
        duration_seconds:
          type: [integer, 'null']
          description: |
            Video categories: send when the client can read it (enables fail-fast UPL-003
            before bytes move); null is always allowed — the server-side probe after upload
            is the authoritative UPL-003 enforcement (CCR-002).
    FileUploadTarget:
      type: object
      required: [file_id, upload_url, method, expires_at]
      properties:
        file_id: { type: string, format: uuid }
        upload_url: { type: string, description: Provider-agnostic presigned target }
        method: { type: string, enum: [PUT, POST] }
        headers:
          type: object
          additionalProperties: { type: string }
        expires_at: { type: string, format: date-time }
    FileUploadComplete:
      type: object
      description: |
        Widened by CCR-004 (v2.0.0 — BREAKING, see the file header note on why a
        major bump was correct here): the original v1.0.0 shape assumed a
        server-side upload session remembering initiate's declared metadata, which
        this API doesn't implement. The client resends content_type/byte_size/
        original_filename/duration_seconds alongside the checksum so complete can
        rebuild the same storage key and re-validate. See docs/reviews/CCR-004-*.md
        for the accepted limitation this carries (no server-side re-read of the
        actual uploaded bytes yet).
      required: [checksum_sha256, content_type, byte_size, original_filename]
      properties:
        checksum_sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
          description: Must equal the initiate value (UPL-005)
        content_type:
          type: string
          description: Must equal the initiate value — re-validated against the category (CCR-004)
        byte_size:
          type: integer
          minimum: 1
        original_filename:
          type: string
          maxLength: 255
        duration_seconds:
          type: [integer, 'null']
          description: Same nullable/probe-authoritative semantics as FileUploadInitiate (CCR-002)
    Mapping:
      type: object
      required: [id, evidence_id, indicator_id, mapping_source, status, mapped_at]
      properties:
        id: { type: string, format: uuid }
        evidence_id: { type: string, format: uuid }
        indicator_id: { type: string, format: uuid }
        cycle_id: { type: [string, 'null'], format: uuid }
        mapping_source: { $ref: '#/components/schemas/MappingSource' }
        status: { $ref: '#/components/schemas/MappingStatus' }
        rationale: { type: [string, 'null'] }
        mapped_by_user_id: { type: string, format: uuid }
        confirmed_by_user_id: { type: [string, 'null'], format: uuid }
        mapped_at: { type: string, format: date-time }
        confirmed_at: { type: [string, 'null'], format: date-time }
...
    PageMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    EvidencePage:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Evidence' }
        meta: { $ref: '#/components/schemas/PageMeta' }
...